scons: verprint() is now called from a proper Command Node to build ver.txt dependency (Copied from rel/callipygous r8031)
The main reason for this change is to fix the strange error
"Implicit dependency `ver.txt' not found, needed by target `build/savegame.rbas.o'"
which occurred the first time scons was run in a clean checkout, or after
running scons -c; rm ver.txt (in that order). I still don't know why that error
occurred, and suspect it's a scons bug, because it wouldn't happen if you print
the ver.txt Node.

Anyway, there are other advantages:
-running verprint is quite slow, and now scons only does so exactly when it's
 actually required, e.g. not when building unlump
-scons -c now also deletes ver.txt, iver.txt, distver.bat
-I removed the list of standard and library includes from ohrbuild.py and use
 a better way of ignoring them as dependencies.




git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8032 7d344553-34f0-0310-a9b1-970ce8f1c3a2