C+1: merge r7997: Add safety checks to killdir, isdir, findfiles after almost wiping my harddisk!
isdir() now returns false for "", findfiles won't return the contents
of / if you pass "", and killdir won't recursively delete / (note that
it continues on error!).
The fix to isdir is significant because it catches lots of possible bugs,
for example, it prevents get_tmpdir() from considering "" a valid tmpdir
and handshake_with_master() from accepting "" as workingdir.



git-svn-id: https://rpg.hamsterrepublic.com/source/rel/callipygous@8049 7d344553-34f0-0310-a9b1-970ce8f1c3a2