Always sort the hero slices to break Y position ties with the leader on top, etc
r8062 did this only when the NPCs and heroes are on the same walkabout layer,
causing a slight inconsistency between different layering modes.

This may cause backcompat issues if anyone ever places multiple heroes on the
same tile and then sorts one to the front, though that seems a very unlikely
thing to want to do, and r8062 would already break this in the default layering
case.

Also, do this sorting in update_walkabout_hero_slices instead of
update_walkabout_npc_slices

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8070 7d344553-34f0-0310-a9b1-970ce8f1c3a2