Wait for a mutex in debuginfo, to be sure that writing to debug log from threads is OK.
This mutex probably isn't needed because FB grabs mutexes in most builtin
functions anyway, and on Unix writes to files opened for append (O_APPEND or "a"
to fopen(), which is what FOR APPEND translates to) can can safely intermingled.
But maybe not on other platforms.

Did not bother to grab it in start/end_debug, although really it's probably
more important there.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8093 7d344553-34f0-0310-a9b1-970ce8f1c3a2