gfx_directx: Mouse2::setPosition() and updatePosition() now handle any way the display could be stretched
Previously, stretching the window with or without aspect preservation,
maximising the window, or setting the zoom to 1x would all cause a call to
setmouse to not necessarily result in the same coord to be reported by getmouse.
In particular, when there are black bars around the edges, the position was
completely wrong.

This rewrite also updates these functions to not assume 320x200.

Note that if the mouse cursor is on the client area (in the window) but outside
the image rect, the reported position is updated, but clamped to within 320x200,
which is a bit inconsistent, but I'll fix that separately.

However, the translation to screen positions by mouse clipping is still wrong.
Also, setPosition() still doesn't truly move the mouse; it'll jump back
to original position when the user touches it.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8115 7d344553-34f0-0310-a9b1-970ce8f1c3a2