findfiles: be certain to let go of DIR-internal handle on Windows
DIR keeps a handle to the search, and if that handle isn't closed the directory
can't be deleted.

This may fix some reports of undeleteable playing/working.tmp, but I doubt it,
because I've only heard of undeleteable files, not directories.
It appears that DIR does always close the HANDLE if DIR is called until it
returns "" (contrary to what I thought), but findfiles does not necessarily
do so when searching for directories.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8203 7d344553-34f0-0310-a9b1-970ce8f1c3a2