Calc autosize menu size in init_menu_state, to fix a couple menu glitches caused by r8241
The problem was that r8241 removed manual .spacing assignments, without which
usemenu can't calculate .size when .autosize=YES, resulting in .size=0 for one
tick, which causes .top to get set to .pt, which is bad if .pt doesn't start
as .first (e.g. slice detail editor remembered .pt).

This mess is caused by usemenu not having access to the data (MenuDef/
MenuOptions + standardmenu arguments) necessary to know .size itself.
And now we have a bunch of standardmenu-specific stuff in MenuState that ought
to be in MenuOptions instead, which usemenu really should be given.
As a simpler fix, I made MenuOptions an argument to init_menu_state instead,
to do the first-tick calculation. Only works when init_menu_state is actuallly
used.

And our menu code just keeps getting more difficult to understand; this
commit not excepted. I avoid looking at menus.bas if possible!

This also cleaned up editbitset slightly.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8248 7d344553-34f0-0310-a9b1-970ce8f1c3a2