vectors: remove all use of unsigned ints, and add more error checking.
It turns out that writing 'x <= v_len(array) - 1' didn't work when the
length was 0, because v_len returned an unsigned 0, which underflowed
to INT_MAX. In C, GCC would throw a warning, but FB doesn't warn
although it follows C's semantics! So unsigned integers are extra risky
in FB, and I will avoid them from now on.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8282 7d344553-34f0-0310-a9b1-970ce8f1c3a2