spriteedit: Add Redo; rewrote undo to use a Frame ptr vector.
Allocate up to 16MB for the undo history (hundreds of times more than before,
which was 32KB). I also doubled the size of the map editor undo history to
16MB, to match.

I got rid of the 'is_rotate' argument to writeundospr, since the only good
reason to skip rotate steps was when the undo history was very short.

Also tried to simplify the code a bit by fiddling with the icon colours
(added textcolor_icon)

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8283 7d344553-34f0-0310-a9b1-970ce8f1c3a2