At last, convert sprite editor to Frames!
As before, the spriteset browser is left as it was, and
is the last part of the source still using 4-bit sprite buffers.

In the process I fixed a bug: one of the recent commits started
clipping the sprite buffer after every rotation.

writeundospr no longer clips the sprite buffer, instead
the various tools call spriteedit_clip if appropriate.
The sprite buffer is left in an unclipped state more
frequently now (e.g. after flipping the buffer).
Also, the points at which undo history is saved has changed
in a few cases.

Net 40 lines removed too

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8288 7d344553-34f0-0310-a9b1-970ce8f1c3a2