Revive & rewrite Animation/SpriteSet/SpriteState code
This code was all quite unfinished and never previously worked,
but it does now! So far, only a couple default animation patterns
are provided as examples. However, nothing here is finalised yet.

A SpriteSet is always created when a spriteset is loaded from
file with frame_load. The Frame and SpriteSet have pointers to
each other, and you can use spriteset_load instead of frame_load;
they are the same aside from which half of the pair is returned.

I rewrote most of it in object-oriented style, but because of the
weird relationship between Frame and SpriteSet it is necessary
to use spriteset_{load,unload} functions. (I could replace these
with class methods, but that would be inconsistent with Frame...)

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8298 7d344553-34f0-0310-a9b1-970ce8f1c3a2