Change Surface Palettes from BGRA to RGBA to fix wrong conversion from RGBcolor
The interpretation of the bytes was swapped when given to the Surface
rendering code and swapped again when displayed, hiding the problem on
the quad-render screen, but not in other cases.

Jay picked BGRA order because that's what DirectX uses (and apparently
most hardware too), but right now it's just a nuisance.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8311 7d344553-34f0-0310-a9b1-970ce8f1c3a2