Fix visual garbage behind initial hero rename prompt; gameover & resetgame plotdict improvements
This garbage would be either the title screen, load menu, or, most
notably, the last frame before "reset game" was called.
I don't think this used to be the case, maybe it was introduced when
I did some cleanup of the game.bas init code not long ago.

In the plotdict, mention that gameover does not fade out the screen but
resetgame does. And some other clarifications.

Also added "game over" and "reset game" options to the Testing Scripts
menu in test.rpg.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8321 7d344553-34f0-0310-a9b1-970ce8f1c3a2