Add battle menu item options for caption, color, tags, hide-if-disabled.
These are designed to overlap with MenuDef. It would be good
if battle menu were just special cases of MenuDefs, so all of the MenuDef
data could be edited. But we're a long way off that being possible,
and I guess when battles and menus are both slicified, the battle slices
could be edited instead to get most of the same customisability.

Since this makes it reasonable to put dozens of items in the battle menu,
also some changes to the editor to handle that better. However, the in-game
menu will extend all the way to the bottom of the screen, probably should be
changed.

New data attributes are "color", "caption", "enable_tag1", "enable_tag2",
"hide_disabled"

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8334 7d344553-34f0-0310-a9b1-970ce8f1c3a2