os: extend open_process so it can be used to fire-and-forget commandline programs
Previously, on Windows it would display a console window, and on Unix
calling cleanup_process would wait for the program to quit. So there
was no function to 'fire and forget' a commandline program. Added some
args to change this behaviour.

There really are a heap of differences between the process_* functions
in os_unix and os_windows, so it's better to avoid them and use the
wrappers in util/common instead.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8343 7d344553-34f0-0310-a9b1-970ce8f1c3a2