Remove xstring/xstringright, replacing them with RelPos values
In most cases I used pCentered instead of pCenteredLeft (which forces
the start of the string to be shown) because it rarely seemed necessary.
When everything gets converted to slices it's a pain to support that behaviour
anyway; better to wrap text.
There are a couple places where I did use wrapprint.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8370 7d344553-34f0-0310-a9b1-970ce8f1c3a2