reloadbasic: Allow putting a comment after a _ line continuation character, like FB
This can misrecognise _ inside comments (there was actually one
comment in common.rbas where this happened). But the old code was also
a kludge.

This doesn't match FB's behaviour, which seems to strip comments and
handle _ as it lexes the line, rather than in a pre-processing step.
RB actually treats comments the same way, but handles _ in a
pre-processing step.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8381 7d344553-34f0-0310-a9b1-970ce8f1c3a2