menus: Unify menu size calculation from standardmenu and init_menu_state, doing so in calc_menustate_size
Added MenuOptions.calc_size, which sets MenuState.size to the actual drawn
extent of the menu instead of stretching to the right and bottom.

I really feel that the menu code is a total mess and is badly holding us
back, every time I need to touch it I know I'm going to be wasting hours.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8395 7d344553-34f0-0310-a9b1-970ce8f1c3a2