Add new backdrop browser, supporting larger backdrops stored in .rgfx format
It is unfinished, since it doesn't actually save backdrops, nor does
Ctrl+Backdrop work, but it's otherwise functional. So for now it is hidden.

This menu is a copy of importbmp without too many changes. After
it is finished, importbmp will only be for tilesets. And longer term,
tilesets will also be converted to rgfx.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8400 7d344553-34f0-0310-a9b1-970ce8f1c3a2