Add one-way walls!
Added a special zone, ID 10000, which marks that all the walls on a tile
can be ignored when exiting the tile.

You can also edit this zone list a regular zone, and read/write with scripts.
Honestly the reason I allow editing it is because it's  hard to hide all trace
of it in the zonemap editor, because mapedit_doZoneHinting would need
to be rewritten.

I haven't updated "check wall collision" yet to properly support
running into edge-on one-way walls... how should that work anyway

I plan to also add a second special zone, for enter-only walls, for more
flexibility, though you can already produce quite complicated wall
arrangements with a single bit.
Another thing that should ideally be added is support to mark oneway
walls on the default wall passability.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8426 7d344553-34f0-0310-a9b1-970ce8f1c3a2