Created the animation editor. Measure animation waits in millisec instead of ticks.
The editor is accessed from the (hidden) new spriteset browser.

This lets you edit and play a single animation (hardcoded to hero 1
"attack" currently), and it's quite functional although nothing is
saved. There's no way to select between or create new animations yet.

Animations ought to be mostly framerate independent now.

Also added any_array_remove macro to remove an item from the middle of
a dynamic array. I know that would be used in a number of places

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8428 7d344553-34f0-0310-a9b1-970ce8f1c3a2