Changed mind again, instead add Global Editor Options menu (F9 anywhere in Custom)
I wasn't sure what key to use; I would have used F8 like the Debug Menu
in-game, but zonemap editing is aclready assigned to F7 so I figured that
F8 might get used someday.
Actually, Game needs another menu for backend/IO config settings anyway,
and F9 might be a good key for that.

Switched the volume key back to F2, which only works in a few menus where
it is advertised. F2 is in easy reach. There's no need to have a global
key for the volume when it's available in the menu instead.

I nearly put Save Game in the menu, but it wouldn't save the changes from
the current menu, which is messy.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8434 7d344553-34f0-0310-a9b1-970ce8f1c3a2