Import build/unix/stdc++compat/stdc++compat.cpp from Mozilla, hg commit 2677f3e14506
Contains hacks to allow compatibility with old versions of libstdc++.so.
I opted to use an older version of the file compatible all the way back
to GCC 4.3 (March 2008), although Mozilla recently deleted most of the
hacks to support only GCC 4.6.1 (June 2011).
See also
https://developer.mozilla.org/en-US/Firefox/Linux_compatibility_matrix
and
https://bugzilla.mozilla.org/show_bug.cgi?id=1278456

This code is licensed under MPL-2.0.txt, which is compatible with
the GPL v2, and interestingly doesn't require any copyright notices
to be displayed when distributed in binary form (unlike MIT and BSD),
only for users to be informed of how they can obtain the source code.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8450 7d344553-34f0-0310-a9b1-970ce8f1c3a2