Add "Test Game with GDB" option
Access either from spam menu, or hold down Shift when selecting Test Game
(holding Shift works both in the main menu and F9 menu).

Works on Linux and Windows, but probably not on OSX, because it requires
xterm. Maybe if you were to install X11 on OSX.

Note the following quirks:
-gdb will break as soon as the channel to Custom has been opened
 (while processing commandline options), which is before setmodex,
 so you need to 'c'ontinue to see the window
-on Unix, Ctrl-C in Custom's terminal will immediately kill Game as
 well because it's sent to the whole process group, instead of
 gdb intercepting it, which is quite annoying
-on Windows, Ctrl-C does nothing (whether using this feature or just
 running gdb normally)

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8471 7d344553-34f0-0310-a9b1-970ce8f1c3a2