Use __FB_UNIX__ instead of __UNIX__ and (new) __GNU_LINUX__ instead of __FB_LINUX__
Did you know that we do
 #ifdef __FB_UNIX__
  #define __FB_LINUX__
 #endif
due to problems with FB's headers? I'd forgotten about that.
For that reason, NEVER use __FB_LINUX__, use the new __GNU_LINUX__
macro instead.

Because of this "running on linux" would have returned true on all
Unixes.

Removed __UNIX__, which was a workaround for old FB versions

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8475 7d344553-34f0-0310-a9b1-970ce8f1c3a2