reloadbasic: add #line directives so the line numbers in FB error messages refer to the original file
Hopefully never again will I have to refer to the .rbas.bas file!

I think I covered all cases, but the implementation of this is basically
a kludge. The way reloadbasic.py outputs lines is very messy.

The exception is that if FB throws an error in the middle some
internal code generated by RB which doesn't have an analogue in the
original file, the line numbers won't mean anything.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8503 7d344553-34f0-0310-a9b1-970ce8f1c3a2