Fix stuff done by autorun scripts before the screen fade in not being hidden
For example if an autorun script made the hero invisible, as the screen
faded in the hero would be visible, and disappear when the fade finished.

This was recently introduced in r8353,
"Limit framerate to 90fps and .gif framerate to 30, overrideable with
-giffps cmdline option"
in which I implemented frame skipping. If a frame is skipped and then the
screen fades in, then the skipped page should be shown, but I typo'd
that bit.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8521 7d344553-34f0-0310-a9b1-970ce8f1c3a2