gfx_directx: Add OHRToWstring(), WstringToOHR() for converting to Latin-1/UTF8
Use it in joystick.cpp to avoid using %S, could break if the ANSI codepage is
something strange.

This is the start of UTF-8 support in the backends, although an alternative
would be to use wstrings in the gfx api and do wstring->utf8 conversion in
allmodex wrappers.

Bumped version number to 1.21, but didn't recompile because neglible difference

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8575 7d344553-34f0-0310-a9b1-970ce8f1c3a2