Add "Don't stop music when starting/loading game" general bitset, and a couple related fixes
The new bit only affects maps set to "Music: -same as previous map-".
In the interests of back-compatibility I didn't want to change the
behaviour without a bitset.

The music was always stopped when loading a save during another game,
even if the new map had the same music, but not when loading from
the titlescreen. I fixed this so that if the music is the same it's
not stopped. ("Same as previous" maps will still be silent without
the new bitset though.)

Don't stop sound effects when loading a game. If you used a 'Load'
special menu item, this would result in the Confirm sfx getting cut off.
But it's not ideal that long sfx will continue.

Also, various helpful comments.

Also turn this on by default in new games.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8602 7d344553-34f0-0310-a9b1-970ce8f1c3a2