attacks: Add necessary upgrade code for zzo38's new attack and damage options.
This increases the attack.bin record size unconditionally, but I left the
initialisation of new data fields in upgrade() disabled for now, so that we
can still change the file format if we want to while implementing these
options in the editor.
Because upgrade() doesn't do it, loadattackdata will upgrade the data as it
loads it, so the new options can be read in-battle.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8642 7d344553-34f0-0310-a9b1-970ce8f1c3a2