backdrop/tileset import: Add "Remap transparent color" and "View with background" options; also prompt for background when importing a 1/4/8-bit bmp
It seemed to be confusing that the backdrop color prompt was only displayed
when importing 24/32 bit bmps. It still is skipped if the palette is identical
to the master, or remapping was cancelled. I also made the help pages much
more elaborate.

"Remap transparent color" also added to new backdrop browser, along with a help
page.

Also renamed "Backdrop:" option to "View with background:" everwhere else

Also fixed fgtag().

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8699 7d344553-34f0-0310-a9b1-970ce8f1c3a2