When importing a backdrop/tileset, don't map to color 0; prompt user for background instead.
Color 0 isn't used unless the palettes are identical or you choose to
not remap (not even if you imported with a new master palette).

This is both for consistency, and in case you have a .bmp which uses
color 0 which you don't want to be mapped to 0: previously, you could
only map an additional color to 0, not prevent the existing on from being used.

Also, in the image color picker, use shift instead of alt to move fast,
and show what the color is getting mapped to.

Also cleanup importbmp_processbmp

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8703 7d344553-34f0-0310-a9b1-970ce8f1c3a2