Add post_event(), called from backends. Add eventFullscreened and user_toggled_fullscreen global
post_termination_signal() is redundant to post_event(), but I'll keep it
since it's still used by the backends.

Got rid of gam.user_toggled_fullscreen and variables of the same name in
the graphics backends; instead added a global to allmodex. Also removed
WindowState.user_toggled_fullscreen.

This is a very minor fix to rungame: now when a new .rpg is started it'sactually
possible to entirely clear the user_toggled_fullscreen flag (previously it
wasn't possible to clear the flags in the backends), which caused gfx.fullscreen
to be unnecessarily saved.

gfx_directx.dll will refuse to run with old versions of game/custom which don't
provide a pointer to post_event. Bumped version to 1.22

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8721 7d344553-34f0-0310-a9b1-970ce8f1c3a2