Fix deletehero(-1) and other negative values deleting hero 0.
deletehero bounded hero IDs to 0...59; I deleted the bound.
I ran scanscripts.py and found there are about 4 games which incorrectly call
 deletehero(findhero(X))
which would delete hero 0 if X isn't in the party, and also break if the
party isn't heroes 0,1,2... in order.
Since X != 0, it's hard to imagine that those games depend on hero 0 being
deleted.  There are also various games that pass a variable to deletehero, which
scanscripts.py can't tell anything about. Nothing else suspicious to be seen.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8734 7d344553-34f0-0310-a9b1-970ce8f1c3a2