upgrade: ensure record counts are correct before doing anything else
There was a lot of code in upgrade that used constants like genMaxTextbox
before they had been verified. This could possibly have resulted in a lot of
blank records getting appended when upgrading, although I didn't see it happen.

This is a pretty significant change to the way upgrade() works, and means that
now any code that initialises a new lump has to set the record count for it too,
so I spent a lot of time checking all that.

Also added fix_record_count lines for genMaxSong and genMaxHero,
and make fix_record_count set the count to zero if the file is missing.

This also fixes up to 2000 error messages from LoadTextBox/SaveTextBox about
invalid record numbers in old games!

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8741 7d344553-34f0-0310-a9b1-970ce8f1c3a2