hspeak: Disallow extra commas in most positions
Extra commas are still allowed before ( and after ), meaning "wait,(1)"
and "if,(foo),then,(bar)" aren't disallowed.
Those are far harder to catch; in fact impossible in the case of "foo(wait,(1))".
Also in the case of if/then it's not uncommon to see extra commas, and arguably
ok.

Also renamed LEX_KEYWORD to LEX_OPERATOR

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8751 7d344553-34f0-0310-a9b1-970ce8f1c3a2