menus: Always pass menu item extra data to a triggered script.
Unfortunately for backcompat the argument order changes based on menu and item
bits: the script gets either 3 or 4 args. But this is more consistent than how
it used to be!

To clear up this very confusing mistake, preview which args get passed to the
script in the menu editor. I want to do this with all script triggers.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8764 7d344553-34f0-0310-a9b1-970ce8f1c3a2