A couple simple optimisations, which have an easily measurable effect on pathfinding
This will change the spacing of the party when the leader's speed doesn't
divide into 20, but if that's the case everything will definitely screw up
anyway. We'll revisit this when other speeds are supported.

catindex is slow, but is never called often enough for it to matter...
except when doing lots of NPC collision tests.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8843 7d344553-34f0-0310-a9b1-970ce8f1c3a2