Add switch_to_{8,32}bit_vpages, and combine allocatepage and allocatepage32
These switch all the video pages to 8 or 32 bits (ie Surface-backed), which
is necessary because you can't draw a Surface-backed Frame onto a videopage
Frame that isn't.
Converting the video pages to 32 bit means that all graphics drawn to them
have to go through an extra layer of remapping as they're drawn, and scaling
up the video page in setvispage becomes slower as well. So it may be too slow
to run a whole game in 32-bit depth

My first was to add a global vpage32 which could be used whenever needed,
but that would mean various functions like multichoice which assume vpage
is used would no longer work.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8895 7d344553-34f0-0310-a9b1-970ce8f1c3a2