gfx_present is now always used so that palette and screen are updated at same time; gfx_showpage is obsolete and no longer used.
This was intended to fix around some wrong-palette glitches seen on Mac
(but which don't happen on linux), such as when opening vikings.rpg in
Custom the screen flashes purple (colour 0) for an instant, and momentary
wrong palette when using "run game".
The purple flash used to occur on Linux but I think was hidden by frame
skipping.
On the other hand frameskipping is broken on Mac, so maybe that's why
it's visible there.

Also, I implemented gfx_present() in gfx_alleg, but can't test because
it doesn't compile for me (mismatched library and headers I think)

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8938 7d344553-34f0-0310-a9b1-970ce8f1c3a2