Add editorconfig.ini file, with one option: show_quit_msg (defaults to true).
There's no editor for this option; you need to edit the file manually.

The search locations:
- .rpg location
- EXEPATH
- settings_dir
If a setting is missing from the foremost file is is NOT searched for in
the next file, but we will probably want that.

The editor_config_file global in Custom points to the file.
(Note there is a config_file global in Game, specific to the current game.)

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8947 7d344553-34f0-0310-a9b1-970ce8f1c3a2