Reenable and rewrite mouse support in menus.
(Yikes, it's been 4 years since James disabled his first attempt!)
The difference is that this time, hovering the mouse over a menu item
highlights it (setting .hover) instead of selecting it.

This is blanket-disabled in Game. I'm not sure whether we want the mouse control
in Game to be per-MenuDef or per-MenuState or global (and overridable when
in a debug menu, etc), and how that should work.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8963 7d344553-34f0-0310-a9b1-970ce8f1c3a2