Rearrange and rename Custom's main menu and Edit Graphics menu items. Moved Edit Font under Edit Graphics
Will this be an annoying change for some? Most items are in the same place
or one spot away from their old location, so I don't think muscle memory
will be too damaged.

A number of menus had weird names. For example, it's odd that the tileset
editor wasn't called that. And "Edit Maps" is better than "Edit Map Data".

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8975 7d344553-34f0-0310-a9b1-970ce8f1c3a2