Fix ellipse slice border/fill colour 0 appearing as uiBackground instead of transparent, if uilook(uiBackground) was not 0
If uilook(uiBackground) is 0 then it acts as transparent when used as
border/fill. To solve that would either require using a mask, or drawing the
ellipse directly to the videopage. But if we switch to hardware acceleration,
then the way we're drawing ellipses (onto a temporary Frame) is already the
correct way to do it, and drawing to a videopage won't be supported.  On the
other hand drawing directly would speed up large ellipses and slow down small
ones.

This bug was introduced in r6519 (in 2014).

Also fixed the plotdict not mentioning that border colour 0 is transparent.
Which is weird, but that's how it's always worked, and it's not easy to change
(see above).

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@8990 7d344553-34f0-0310-a9b1-970ce8f1c3a2