Textbox editor: create and link to a new textbox by pressing Enter, + or Insert on the "After:" field
Pressing Enter only works if After is None; it brings up a menu asking what
you want.

There are now 8 ways to add a new textbox, if you include importing a .txt
and pressing +/Insert on another menu as one each.

Maybe it's overkill to show the "+ or INSERT to link to new text box"
message when that menu option is selected. I nearly left it out. But I think
a lot of people won't realise you can do this otherwise.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9057 7d344553-34f0-0310-a9b1-970ce8f1c3a2