Add F1 Textbox debug info debug key
Only works when a textbox is up.
It was already the case that debug keys F1-F3 weren't available when a textbox
was up. That was because saving/loading during a textbox may cause problems.
There wasn't a good reason to disallow F1/Ctrl-F1 though. But they're not very
useful during a textbox anyway.
Pressing F1 to see info about a textbox feels very natural, but I'm not against
changing it

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9089 7d344553-34f0-0310-a9b1-970ce8f1c3a2