Regretfully switch to dynamically loading libvorbisfile at runtime.
I'm having some trouble with building on OSX in a way that works both with
and without bundling the .apps.

Maybe all this loading code can be removed in future, but it does have
the advantage that the engine can work if libvorbisfile is unexpectantly
missing at runtime. That shouldn't really happen, except when someone uses
an old copy of SDL_mixer.dll

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9118 7d344553-34f0-0310-a9b1-970ce8f1c3a2