Add a modest object-orientated generic menu system, ModularMenu
This does not compete with standardmenu/MenuDef, but instead wraps them.
It competes with multichoice() and sublist().
Right now what this accomplishes is reducing the boiler plate of the
same loop around standardmenu everywhere.
But hopefully in future this moves more towards declarative programming,
defining menu items cleanly.
Probably this can be merge with editrunner.

It has a 'floating' mode where it acts more like a modal multichoice() menu
instead of a fullscreen one, but it doesn't look or work too well. Should
use MenuDef instead of standardmenu.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9124 7d344553-34f0-0310-a9b1-970ce8f1c3a2