sliceedit: workaround for when both a slice and a menu item are under the mouse
Clicking now selects the slice. If there's no selectable slice, then the menu
item is activated instead. Right clicking can also be used to select (not
activate) a menu item.

This doesn't feel very good (maybe it should do the opposite, with right clicking to select a slice, and left clicking always activating a menu item if any),
but the previous behaviour was completely broken

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9210 7d344553-34f0-0310-a9b1-970ce8f1c3a2