Fix 'Follow Walls Stop For Others' incorrectly turning when bumping into someone while rounding a corner.
This movetype attempts to turn and then move forward a tile when it reaches a
corner, but the movement might fail if an npc/hero is there. In that case, next
tick it needs to know that it's already turned, and not try to turn again until
it completes that movement. I added a bool to NPCInst to track that.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9227 7d344553-34f0-0310-a9b1-970ce8f1c3a2