Lift most of the stat limits in the hero/enemy/stat caps menus to 32767
There are no limits when using scripting to set stats. The editor caps
were often even 999 instead of 9999, and were inconsistent between the
3 editors.
I kept Focus limited at 100 (because I'm not sure how Focus >100 should
affect spells with negative MP costs: should be negate to positive?)
and raised the Hits cap to 20, because that's what it was in the stat
caps menu.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@9331 7d344553-34f0-0310-a9b1-970ce8f1c3a2